ButtonGroup: Convert to TypeScript
