+++ packages/components/src/button-group/types.ts
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import type { ReactNode } from 'react';
+
+export type ButtonGroupProps = {
+	/**
+	 * The children elements.
+	 */
+	children: ReactNode;
+};
